+++ lib/src/native/realm_core.dart
@@ -1751,61 +1751,42 @@ class _RealmCore {
     return completer.future;
   }
 
-  static void _openRealmAsyncCallback(Handle userdata, Pointer<realm_thread_safe_reference_t> realm, Pointer<realm_async_error> error) {
-    final completer = userdata as Completer<RealmHandle?>;
-    if (error != nullptr) {
-      using((Arena arena) {
-        final out_error = arena<realm_error>();
-        _realmLib.realm_get_async_error(error, out_error);
-        final message = out_error.ref.message.cast<Utf8>().toRealmDartString();
-        completer.completeError(RealmException(message!));
-      });
-    } else {
-      final realmPtr =
-          _realmLib.invokeGetPointer(() => _realmLib.realm_from_thread_safe_reference(realm, scheduler.handle._pointer), "Error opening synchronized realm");
-      if (completer.isCompleted) {
-        //Task is canceled before to complete but the realm was created.
-        _realmLib.invokeGetBool(() => _realmLib.realm_close(realmPtr), "Realm close failed");
-      } else {
-        completer.complete(RealmHandle._(realmPtr));
+  bool isFrozen(Realm realm) {
+    return _realmLib.realm_is_frozen(realm.handle._pointer.cast());
+  }
+
+  RealmHandle freeze(Realm realm) {
+    final ptr = _realmLib.invokeGetPointer(() => _realmLib.realm_freeze(realm.handle._pointer));
+    return RealmHandle._(ptr);
+  }
+
+  RealmResultsHandle resolveResults(RealmResults realmResults, Realm frozenRealm) {
+    final ptr = _realmLib.invokeGetPointer(() => _realmLib.realm_results_resolve_in(realmResults.handle._pointer, frozenRealm.handle._pointer));
+    return RealmResultsHandle._(ptr);
+  }
+
+  RealmObjectHandle? resolveObject(RealmObject object, Realm frozenRealm) {
+    return using((Arena arena) {
+      final resultPtr = arena<Pointer<realm_object>>();
+      _realmLib.invokeGetBool(() => _realmLib.realm_object_resolve_in(object.handle._pointer, frozenRealm.handle._pointer, resultPtr));
+      if (resultPtr != nullptr) {
+        return RealmObjectHandle._(resultPtr.value);
       }
-    }
+
+      return null;
+    });
   }
 
-  Future<RealmHandle?> openRealmAsync(RealmAsyncOpenTaskHandle handle, Completer<RealmHandle?> completer) {
-    final callback = Pointer.fromFunction<Void Function(Handle, Pointer<realm_thread_safe_reference_t>, Pointer<realm_async_error_t>)>(_openRealmAsyncCallback);
-    final userdata = _realmLib.realm_dart_userdata_async_new(completer, callback.cast(), scheduler.handle._pointer);
+  RealmListHandle? resolveList(ManagedRealmList list, Realm frozenRealm) {
+    return using((Arena arena) {
+      final resultPtr = arena<Pointer<realm_list>>();
+      _realmLib.invokeGetBool(() => _realmLib.realm_list_resolve_in(list.handle._pointer, frozenRealm.handle._pointer, resultPtr));
+      if (resultPtr != nullptr) {
+        return RealmListHandle._(resultPtr.value);
+      }
 
-    _realmLib.realm_async_open_task_start(handle._pointer, _realmLib.addresses.realm_dart_async_open_task_completion_callback, userdata.cast(),
-        _realmLib.addresses.realm_dart_userdata_async_free);
-    return completer.future;
-  }
-
-  RealmAsyncOpenTaskHandle createRealmAsyncOpenTask(Configuration config) {
-    final configHandle = _createConfig(config);
-    final realmAsyncOpenTaskPtr =
-        _realmLib.invokeGetPointer(() => _realmLib.realm_open_synchronized(configHandle._pointer), "Error opening synchronized realm at path ${config.path}");
-    return RealmAsyncOpenTaskHandle._(realmAsyncOpenTaskPtr);
-  }
-
-  void cancelRealmAsyncOpenTask(RealmAsyncOpenTaskHandle handle) {
-    _realmLib.realm_async_open_task_cancel(handle._pointer);
-  }
-
-  static void _realmAsyncOpenRegisterProgressNotifierCallback(Handle userdata, int transferredBytes, int totalBytes) {
-    final progressCallback = userdata as void Function(int transferredBytes, int totalBytes);
-    progressCallback(transferredBytes, totalBytes);
-  }
-
-  int realmAsyncOpenRegisterProgressNotifier(RealmAsyncOpenTaskHandle handle, void Function(int transferredBytes, int totalBytes) progressCallback) {
-    final callback = Pointer.fromFunction<Void Function(Handle, Uint64, Uint64)>(_realmAsyncOpenRegisterProgressNotifierCallback);
-    final userdata = _realmLib.realm_dart_userdata_async_new(progressCallback, callback.cast(), scheduler.handle._pointer);
-    return _realmLib.realm_async_open_task_register_download_progress_notifier(
-        handle._pointer, _realmLib.addresses.realm_dart_sync_progress_callback, userdata.cast(), _realmLib.addresses.realm_dart_userdata_async_free);
-  }
-
-  void realmAsyncOpenUnregisterProgressNotifier(RealmAsyncOpenTaskHandle handle, int token) {
-    _realmLib.realm_async_open_task_unregister_download_progress_notifier(handle._pointer, token);
+      return null;
+    });
   }
 }
 
@@ -1879,10 +1860,6 @@ class RealmHandle extends HandleBase<shared_realm> {
   RealmHandle._(Pointer<shared_realm> pointer) : super(pointer, 24);
 
   RealmHandle._unowned(Pointer<shared_realm> pointer) : super.unowned(pointer);
-}
-
-class RealmAsyncOpenTaskHandle extends HandleBase<realm_async_open_task> {
-  RealmAsyncOpenTaskHandle._(Pointer<realm_async_open_task> pointer) : super(pointer, 24);
 }
 
 class SchedulerHandle extends HandleBase<realm_scheduler> {