+++ src/realm_dart_sync.h
@@ -35,5 +35,3 @@ RLM_API void realm_dart_sync_connection_state_changed_callback(realm_userdata_t 
                                                                realm_sync_connection_state_e new_state);
 
 RLM_API void realm_dart_sync_on_subscription_state_changed_callback(realm_userdata_t userdata, realm_flx_sync_subscription_set_state_e state);
-
-RLM_API void realm_dart_async_open_task_completion_callback(realm_userdata_t userdata, realm_thread_safe_reference_t* realm, realm_async_error_t* error);